fix: add temporary fix for high compression ratio files
Currently, we have a hardcoded buffer size of compressed * 4 for the
decompressed data, since deflate does not give the true decompressed
size (although zip metadata does). So, on files with a very high
compression ratio, a buffer overflow is triggered as our destination
buffer is not large enough for the entire decompressed data.

This temporary fix involves increasing the factor from 4 to 7 for the
size estimation. A future fix would involve using the true decompressed
size.