chore(actions): fix `tty` device being nonexistent in test step
`/dev/tty` does not exist in an actions step since it is being run non
interactively. Similarly using the stdout device does not work either,
since we're within a command substitution expression, where `/dev/stdout`
is not attached to a shell.

Instead, we have to use `/dev/stderr`, which is attached to a shell at
all times. I don't know if there are any better alternatives, but
writing to stderr doesn't sit quite right with me, since this should
ideally output to stdout. Oh, well.